feat: per-user rate limit multiplier with admin endpoints
